Update accounts_ce.py for lava output and merge accounts_ce_authtokens.py
